check if core component exists.

+++ admin/joomla/1.5/trunk/com_easysdi_core/uninstall.easysdi.php
@@ -21,15 +21,28 @@ defined('_JEXEC') or die('Restricted access');
 function com_uninstall(){
 	
 global  $mainframe;
-
 $db =& JFactory::getDBO();
+
+
+$query = "select count(*) FROM #__components where `option` in ('com_easysdi_proxy','com_easysdi_catalog','com_easysdi_shop' ";
+$db->setQuery( $query);
+$total = $db->loadResult();
+if($total >0){
+	
+	$mainframe->enqueueMessage("core component could not be uninstalled, because some depending components are still installed. Please uninstall all others easysdi component before uninstalling this component.","ERROR");	
+	return false;
+}
+
+
+
+
 $query = "DELETE FROM #__components where `option`= 'com_easysdi_partner' ";
 
 $db->setQuery( $query);
 
 	if (!$db->query()) {
 		$mainframe->enqueueMessage($db->getErrorMsg(),"ERROR");
-		
+		return false;
 	}
 
 	$query = "DELETE FROM #__components where `option`= 'com_easysdi_core' ";
@@ -38,7 +51,7 @@ $db->setQuery( $query);
 
 	if (!$db->query()) {
 		$mainframe->enqueueMessage($db->getErrorMsg(),"ERROR");
-		
+		return false;
 	}
 	
 $mainframe->enqueueMessage("Congratulation EasySdi core component is uninstalled.
